+++ tests/importers/__init__.py
@@ -1,0 +1,3 @@
+"""coBib's importer tests."""
+
+from .importer_test import ImporterTest